+++ plugin/intellij/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- include licenses in bundle
+
 ### Fixed
 
 - update bundled docsig from v0.87.1 to v0.90.3 for several false