+++ plugin/intellij/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.0] - 2026-07-15
+
 ### Changed
 
 - include licenses in bundle
@@ -42,7 +44,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial JetBrains Marketplace release with bundled docsig checker
   (docsig v0.86.1)
 
-[Unreleased]: https://github.com/jshwi/docsig/compare/1.0.2...HEAD
+[Unreleased]: https://github.com/jshwi/docsig/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/jshwi/docsig/compare/1.0.2...1.1.0
 [1.0.2]: https://github.com/jshwi/docsig/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/jshwi/docsig/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/jshwi/docsig/commits/1.0.0